Reuse single Pipeline type in Client, ClusterClient and Ring.
